Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 
 int main() {
-    int a,b,c;
-    scanf("%d", &a);
-    scanf("%d", &b);
-    scanf("%d", &c);
+    float a,b,c;
+    scanf("%f", &a);
+    scanf("%f", &b);
+    scanf("%f", &c);
     int average = (a+b+c)/3;
-    printf("Average: %d", average);
+    printf("Average: %f", average);
     return 0;
 }
